pyproject.toml
[project]
name = "drf-auth"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1.6",
    "djangorestframework>=3.15.2",
    "pyjwt>=2.10.1",
    "ruff>=0.9.9",
]

[tool.ruff]
line-length = 100  
target-version = "py311"

[tool.ruff.lint]
select = ["ALL"]  
ignore = [
    "E501", "D", "UP035", "ANN401",
    "B010", "B009", "SIM117", "EM101",
    "TRY003", "B904"
]  
